Uploaded Machines JSON Template
Useful template for creating machines json for Content Patcher mods for Stardew Valley. 
All comments and notes are my own and may contain errors and/or may be out of date depending on Content Patcher version - always validate against the Content Patcher documentation or the Modding Wiki in case of any doubts.